+++ src/renderer/containers/JobFailureDialog.tsx
@@ -1,7 +1,11 @@
 import * as React from 'react';
 import { ModalDialog } from "../components/ModalDialog";
-import { ERROR_CODE_INVALID_PARAMS, ERROR_CODE_OS_ERROR, ERROR_CODE_OUT_OF_MEMORY, JobFailure } from "../webapi";
-import { Button, Checkbox, Collapse, Label } from "@blueprintjs/core";
+import {
+    ERROR_CODE_INVALID_PARAMS,
+    getJobFailureIconName, getJobFailureTitle,
+    JobFailure
+} from "../webapi";
+import { Button, Checkbox, Collapse, IconName, Label } from "@blueprintjs/core";
 import * as selectors from "../selectors";
 import { State } from "../state";
 import { connect, DispatchProp } from "react-redux";
@@ -80,18 +84,11 @@ class JobFailureDialog extends React.Component<DispatchProp<State> & IJobFailure
         if (!this.props.isOpen || !this.props.jobFailure) {
             return null;
         }
-        let title;
-        if (this.props.jobFailure.code === ERROR_CODE_INVALID_PARAMS) {
-            title = 'Input Validation Failed';
-        } else if (this.props.jobFailure.code === ERROR_CODE_OS_ERROR) {
-            title = 'Operation System Error';
-        } else if (this.props.jobFailure.code === ERROR_CODE_OUT_OF_MEMORY) {
-            title = 'Out-Of-Memory Error';
-        } else {
-            title = 'Error';
-        }
+        const iconName = getJobFailureIconName(this.props.jobFailure) as IconName;
+        const title = getJobFailureTitle(this.props.jobFailure);
         return (
             <ModalDialog isOpen={this.props.isOpen}
+                         iconName={iconName}
                          title={title}
                          onConfirm={this.handleConfirm}
                          onCancel={this.handleCancel}
@@ -109,7 +106,12 @@ class JobFailureDialog extends React.Component<DispatchProp<State> & IJobFailure
         if (!messageText) {
             messageText = `An unknown error occurred (code ${this.props.jobFailure.code}).`;
         }
-        let message = (<p>{this.props.jobTitle}:<br/><strong>{messageText}</strong></p>);
+        let message = (
+            <div>
+                <p>Oops, Cate couldn't carry out the request <em>{this.props.jobTitle}</em> because:</p>
+                <h6 className="user-selectable">{messageText}</h6>
+            </div>
+        );
 
         let traceback;
         if (this.props.jobFailure.data && this.props.jobFailure.data.traceback) {
@@ -117,7 +119,8 @@ class JobFailureDialog extends React.Component<DispatchProp<State> & IJobFailure
                 <div style={{marginTop: '0.5em'}}>
                     <Button onClick={this.handleShowDetails}>{this.state.showDetails ? "Hide" : "Show"} Details</Button>
                     <Collapse isOpen={this.state.showDetails}>
-                        <pre style={{overflow: 'auto', height: '20em'}}>{this.props.jobFailure.data.traceback}</pre>
+                        <pre className="user-selectable"
+                             style={{overflow: 'auto', height: '20em'}}>{this.props.jobFailure.data.traceback}</pre>
                     </Collapse>
                 </div>
             );
@@ -128,7 +131,7 @@ class JobFailureDialog extends React.Component<DispatchProp<State> & IJobFailure
             reporting = (
                 <Label
                     style={{marginTop: '0.5em'}}
-                    text={<span>Please consider reporting this issue in Cate's <a
+                    text={<span className="pt-text-muted">Please consider reporting this issue in Cate's <a
                         href="https://github.com/CCI-Tools/cate/issues" target="_blank">issue tracker</a>.</span>}
                 >
                     <Checkbox label='Copy error report to clipboard'