+++ src/renderer/containers/PlacemarksPanel.tsx
@@ -17,7 +17,7 @@ import {GeometryToolType} from "../components/cesium/geometry-tool";
 import {isBoolean} from "../../common/types";
 import {NumericField, NumericFieldValue} from "../components/field/NumericField";
 import {ToolButton} from "../components/ToolButton";
-import {CSSProperties} from 'react';
+import { CSSProperties } from "react";
 
 interface IPlacemarksPanelDispatch {
     dispatch: Dispatch<State>;
@@ -355,8 +355,8 @@ interface IPlacemarkItemProps {
 @ContextMenuTarget
 class PlacemarkItem extends React.PureComponent<IPlacemarkItemProps, {}> {
 
-    static readonly ICON_STYLE = {marginLeft: "0.5em"};
-    static readonly NAME_STYLE = {marginLeft: "0.5em"};
+    static readonly ICON_STYLE: CSSProperties = {marginLeft: "0.5em"};
+    static readonly NAME_STYLE: CSSProperties = {marginLeft: "0.5em"};
     static readonly INFO_STYLE: CSSProperties = {float: "right", color: Colors.BLUE5};
 
     constructor(props: IPlacemarkItemProps) {